fix(computedInject): demo page
